Replaced ULN2803 pump driver with MOSFETs
Despite the ULN2803 supposedly having built-in protection for driving
inductive loads, we occasionally observed some ULNs going up in flames
when the pump was switched on and off a lot (despite the current being
well within the rating of the ULN2803).

Initial testing shows that a DMN6070SSD dual N-channel MOSFET makes a
great dual pump driver; therefore, this has replaced the ULN in this
design.